almost nothing

+++ h/Semaphore.hpp
@@ -11,22 +11,22 @@ public:
     void* operator new(size_t n)   { return MemoryAllocator::getInstance().memAlloc(n); }
     void  operator delete(void* p) { MemoryAllocator::getInstance().free(p); }
 
-    _Semaphore(uint64 init);
+    _Semaphore(int init);
     ~_Semaphore();
 
-    static _Semaphore* create_semaphore(uint64 init);
+    static _Semaphore* create_semaphore(int init);
     int getVal() const { return this->val; }
 
     void signal();
     void wait();
 
-    void signal_n(uint64 n);
-    void wait_n(uint64 n);
+    void signal_n(int n);
+    void wait_n(int n);
 
     void block();
     void unblock();
 private:
     List<TCB> blocked;
-    uint64 val;
+    int val;
 };
 #endif
